third_party/blink/web_tests/external/wpt/css/cssom/idlharness-expected.txt

This is a testharness.js-based test.
Found 57 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] CSSRule interface: constant MARGIN_RULE on interface object
  assert_own_property: expected property "MARGIN_RULE" missing
[FAIL] CSSRule interface: constant MARGIN_RULE on interface prototype object
  assert_own_property: expected property "MARGIN_RULE" missing
[FAIL] CSSStyleRule interface: existence and properties of interface object
  assert_equals: prototype of CSSStyleRule is not CSSGroupingRule expected function "function CSSGroupingRule() { [native code] }" but got function "function CSSRule() { [native code] }"
[FAIL] CSSStyleRule interface: existence and properties of interface prototype object
  assert_equals: prototype of CSSStyleRule.prototype is not CSSGroupingRule.prototype expected object "[object CSSGroupingRule]" but got object "[object CSSRule]"
[FAIL] CSSRule interface: sheet.cssRules[4] must inherit property "MARGIN_RULE" with the proper type
  assert_inherits: property "MARGIN_RULE" not found in prototype chain
[FAIL] CSSImportRule interface: sheet.cssRules[0] must inherit property "layerName" with the proper type
  assert_equals: expected "string" but got "object"
[FAIL] CSSImportRule interface: sheet.cssRules[0] must inherit property "supportsText" with the proper type
  assert_equals: expected "string" but got "object"
[FAIL] CSSRule interface: sheet.cssRules[0] must inherit property "MARGIN_RULE" with the proper type
  assert_inherits: property "MARGIN_RULE" not found in prototype chain
[FAIL] CSSPageRule interface: existence and properties of interface object
  assert_equals: prototype of CSSPageRule is not CSSGroupingRule expected function "function CSSGroupingRule() { [native code] }" but got function "function CSSRule() { [native code] }"
[FAIL] CSSPageRule interface: existence and properties of interface prototype object
  assert_equals: prototype of CSSPageRule.prototype is not CSSGroupingRule.prototype expected object "[object CSSGroupingRule]" but got object "[object CSSRule]"
[FAIL] CSSGroupingRule interface: sheet.cssRules[2] must inherit property "cssRules" with the proper type
  assert_inherits: property "cssRules" not found in prototype chain
[FAIL] CSSGroupingRule interface: sheet.cssRules[2] must inherit property "insertRule(CSSOMString, optional unsigned long)" with the proper type
  assert_inherits: property "insertRule" not found in prototype chain
[FAIL] CSSGroupingRule interface: calling insertRule(CSSOMString, optional unsigned long) on sheet.cssRules[2] with too few arguments must throw TypeError
  assert_inherits: property "insertRule" not found in prototype chain
[FAIL] CSSGroupingRule interface: sheet.cssRules[2] must inherit property "deleteRule(unsigned long)" with the proper type
  assert_inherits: property "deleteRule" not found in prototype chain
[FAIL] CSSGroupingRule interface: calling deleteRule(unsigned long) on sheet.cssRules[2] with too few arguments must throw TypeError
  assert_inherits: property "deleteRule" not found in prototype chain
[FAIL] CSSRule interface: sheet.cssRules[2] must inherit property "MARGIN_RULE" with the proper type
  assert_inherits: property "MARGIN_RULE" not found in prototype chain
[FAIL] CSSMarginRule interface: existence and properties of interface object
  assert_own_property: self does not have own property "CSSMarginRule" expected property "CSSMarginRule" missing
[FAIL] CSSMarginRule interface object length
  assert_own_property: self does not have own property "CSSMarginRule" expected property "CSSMarginRule" missing
[FAIL] CSSMarginRule interface object name
  assert_own_property: self does not have own property "CSSMarginRule" expected property "CSSMarginRule" missing
[FAIL] CSSMarginRule interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "CSSMarginRule" expected property "CSSMarginRule" missing
[FAIL] CSSMarginRule interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "CSSMarginRule" expected property "CSSMarginRule" missing
[FAIL] CSSMarginRule interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "CSSMarginRule" expected property "CSSMarginRule" missing
[FAIL] CSSMarginRule interface: attribute name
  assert_own_property: self does not have own property "CSSMarginRule" expected property "CSSMarginRule" missing
[FAIL] CSSMarginRule interface: attribute style
  assert_own_property: self does not have own property "CSSMarginRule" expected property "CSSMarginRule" missing
[FAIL] CSSMarginRule must be primary interface of sheet.cssRules[2].cssRules[0]
  assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Cannot read properties of undefined (reading '0')"
[FAIL] Stringification of sheet.cssRules[2].cssRules[0]
  assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Cannot read properties of undefined (reading '0')"
[FAIL] CSSMarginRule interface: sheet.cssRules[2].cssRules[0] must inherit property "name" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Cannot read properties of undefined (reading '0')"
[FAIL] CSSMarginRule interface: sheet.cssRules[2].cssRules[0] must inherit property "style" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Cannot read properties of undefined (reading '0')"
[FAIL] CSSRule interface: sheet.cssRules[2].cssRules[0] must inherit property "cssText" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Cannot read properties of undefined (reading '0')"
[FAIL] CSSRule interface: sheet.cssRules[2].cssRules[0] must inherit property "parentRule" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Cannot read properties of undefined (reading '0')"
[FAIL] CSSRule interface: sheet.cssRules[2].cssRules[0] must inherit property "parentStyleSheet" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Cannot read properties of undefined (reading '0')"
[FAIL] CSSRule interface: sheet.cssRules[2].cssRules[0] must inherit property "type" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Cannot read properties of undefined (reading '0')"
[FAIL] CSSRule interface: sheet.cssRules[2].cssRules[0] must inherit property "STYLE_RULE" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Cannot read properties of undefined (reading '0')"
[FAIL] CSSRule interface: sheet.cssRules[2].cssRules[0] must inherit property "CHARSET_RULE" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Cannot read properties of undefined (reading '0')"
[FAIL] CSSRule interface: sheet.cssRules[2].cssRules[0] must inherit property "IMPORT_RULE" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Cannot read properties of undefined (reading '0')"
[FAIL] CSSRule interface: sheet.cssRules[2].cssRules[0] must inherit property "MEDIA_RULE" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Cannot read properties of undefined (reading '0')"
[FAIL] CSSRule interface: sheet.cssRules[2].cssRules[0] must inherit property "FONT_FACE_RULE" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Cannot read properties of undefined (reading '0')"
[FAIL] CSSRule interface: sheet.cssRules[2].cssRules[0] must inherit property "PAGE_RULE" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Cannot read properties of undefined (reading '0')"
[FAIL] CSSRule interface: sheet.cssRules[2].cssRules[0] must inherit property "MARGIN_RULE" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Cannot read properties of undefined (reading '0')"
[FAIL] CSSRule interface: sheet.cssRules[2].cssRules[0] must inherit property "NAMESPACE_RULE" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Cannot read properties of undefined (reading '0')"
[FAIL] CSSRule interface: sheet.cssRules[1] must inherit property "MARGIN_RULE" with the proper type
  assert_inherits: property "MARGIN_RULE" not found in prototype chain
[FAIL] CSSStyleDeclaration must be primary interface of sheet.cssRules[2].cssRules[0].style
  assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Cannot read properties of undefined (reading '0')"
[FAIL] Stringification of sheet.cssRules[2].cssRules[0].style
  assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Cannot read properties of undefined (reading '0')"
[FAIL] CSSStyleDeclaration interface: sheet.cssRules[2].cssRules[0].style must inherit property "cssText" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Cannot read properties of undefined (reading '0')"
[FAIL] CSSStyleDeclaration interface: sheet.cssRules[2].cssRules[0].style must inherit property "length" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Cannot read properties of undefined (reading '0')"
[FAIL] CSSStyleDeclaration interface: sheet.cssRules[2].cssRules[0].style must inherit property "item(unsigned long)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Cannot read properties of undefined (reading '0')"
[FAIL] CSSStyleDeclaration interface: calling item(unsigned long) on sheet.cssRules[2].cssRules[0].style with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Cannot read properties of undefined (reading '0')"
[FAIL] CSSStyleDeclaration interface: sheet.cssRules[2].cssRules[0].style must inherit property "getPropertyValue(CSSOMString)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Cannot read properties of undefined (reading '0')"
[FAIL] CSSStyleDeclaration interface: calling getPropertyValue(CSSOMString) on sheet.cssRules[2].cssRules[0].style with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Cannot read properties of undefined (reading '0')"
[FAIL] CSSStyleDeclaration interface: sheet.cssRules[2].cssRules[0].style must inherit property "getPropertyPriority(CSSOMString)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Cannot read properties of undefined (reading '0')"
[FAIL] CSSStyleDeclaration interface: calling getPropertyPriority(CSSOMString) on sheet.cssRules[2].cssRules[0].style with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Cannot read properties of undefined (reading '0')"
[FAIL] CSSStyleDeclaration interface: sheet.cssRules[2].cssRules[0].style must inherit property "setProperty(CSSOMString, CSSOMString, optional CSSOMString)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Cannot read properties of undefined (reading '0')"
[FAIL] CSSStyleDeclaration interface: calling setProperty(CSSOMString, CSSOMString, optional CSSOMString) on sheet.cssRules[2].cssRules[0].style with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Cannot read properties of undefined (reading '0')"
[FAIL] CSSStyleDeclaration interface: sheet.cssRules[2].cssRules[0].style must inherit property "removeProperty(CSSOMString)" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Cannot read properties of undefined (reading '0')"
[FAIL] CSSStyleDeclaration interface: calling removeProperty(CSSOMString) on sheet.cssRules[2].cssRules[0].style with too few arguments must throw TypeError
  assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Cannot read properties of undefined (reading '0')"
[FAIL] CSSStyleDeclaration interface: sheet.cssRules[2].cssRules[0].style must inherit property "parentRule" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Cannot read properties of undefined (reading '0')"
[FAIL] CSSStyleDeclaration interface: sheet.cssRules[2].cssRules[0].style must inherit property "cssFloat" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Cannot read properties of undefined (reading '0')"
Harness: the test ran to completion.
